Harden and refactor object data updates
