+++ src/api/v1/handlers/classes.rs
@@ -42,11 +42,11 @@ use crate::models::{
     ClassGraphRow, ClassSelector, CollectionID, GroupPermission, HubuumClass, HubuumClassExpanded,
     HubuumClassHistory, HubuumClassID, HubuumClassRelation, HubuumClassRelationID,
     HubuumClassWithPath, HubuumObject, HubuumObjectHistory, HubuumObjectID,
-    HubuumObjectReadResponse, HubuumObjectRelation, HubuumObjectWithPath,
-    NewHubuumClass, NewHubuumClassRelationFromClass, NewHubuumObjectRelation,
-    NewHubuumObjectRequest, ObjectDataPatchDocument, ObjectSelector, Permissions,
-    RelatedClassGraph, RelatedObjectGraph, RelatedObjectGraphRow, ResolvedClassTarget,
-    ResolvedObjectTarget, UpdateHubuumClass, UpdateHubuumObject, UpdateHubuumObjectRequest,
+    HubuumObjectReadResponse, HubuumObjectRelation, HubuumObjectWithPath, NewHubuumClass,
+    NewHubuumClassRelationFromClass, NewHubuumObjectRelation, NewHubuumObjectRequest,
+    ObjectDataPatchDocument, ObjectSelector, Permissions, RelatedClassGraph, RelatedObjectGraph,
+    RelatedObjectGraphRow, ResolvedClassTarget, ResolvedObjectTarget, UpdateHubuumClass,
+    UpdateHubuumObject, UpdateHubuumObjectRequest,
 };
 use crate::traits::{BackendContext, CanDelete, CanSave, Search, SelfAccessors};
 use crate::utilities::extensions::CustomStringExtensions;
@@ -58,8 +58,6 @@ use crate::models::search::{
 };
 use crate::models::traits::class_relation::ToHubuumClasses;
 use crate::pagination::{Page, finalize_page};
-
-mod computed_objects;
 
 fn parse_computed_include(query_string: &str) -> Result<(QueryOptions, bool), ApiError> {
     let (params, mut passthrough) =
@@ -91,7 +89,7 @@ fn parse_computed_object_list_query(query_string: &str) -> Result<(QueryOptions,
     Ok((params, include_computed))
 }
 
-pub(super) fn scope_object_query_to_class(params: &mut QueryOptions, class: &HubuumClassID) {
+fn scope_object_query_to_class(params: &mut QueryOptions, class: &HubuumClassID) {
     params
         .filters
         .retain(|param| !matches!(param.field, FilterField::ClassId | FilterField::Classes));
@@ -102,7 +100,7 @@ pub(super) fn scope_object_query_to_class(params: &mut QueryOptions, class: &Hub
     );
 }
 
-pub(super) async fn computed_personal_owner(
+async fn computed_personal_owner(
     pool: &AppContext,
     requestor: &Authenticated,
     class: &HubuumClass,
@@ -787,2090 +785,13 @@ async fn read_resolved_class_permissions(
     ApiResponse::paginated(permissions, total_count, &params)
 }
 
-#[utoipa::path(
-    get,
-    path = "/api/v1/classes/{class_id}/related/classes",
-    tag = "classes",
-    security(("bearer_auth" = [])),
-    params(
-        ("class_id" = i32, Path, description = "Class ID")
-    ),
-    responses(
-        (status = 200, description = "Classes connected to the class", body = [HubuumClassWithPath]),
-        (status = 400, description = "Bad request", body = ApiErrorResponse),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 404, description = "Class not found", body = ApiErrorResponse)
-    )
-)]
-#[routes]
-#[get("/{class_id}/related/classes")]
-#[get("/{class_id}/related/classes/")]
-async fn get_related_classes(
-    pool: AppContext,
-    requestor: Authenticated,
-    class_id: web::Path<HubuumClassID>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    let target = ClassSelector::by_id(class_id.into_inner())
-        .resolve_class_target(&pool)
-        .await?;
-    read_related_classes(pool, requestor, target, req).await
-}
-
-#[utoipa::path(
-    get,
-    path = "/api/v1/classes/by-name/{class_name}/related/classes",
-    tag = "classes",
-    summary = "Get related classes by class name",
-    security(("bearer_auth" = [])),
-    params(("class_name" = String, Path, description = "Globally unique class name")),
-    responses(
-        (status = 200, description = "Classes connected to the class", body = [HubuumClassWithPath]),
-        (status = 400, description = "Bad request", body = ApiErrorResponse),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 404, description = "Class not found", body = ApiErrorResponse)
-    )
-)]
-#[get("/by-name/{class_name}/related/classes")]
-async fn get_related_classes_by_name(
-    pool: AppContext,
-    requestor: Authenticated,
-    class_name: web::Path<String>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    let target = ClassSelector::by_name(class_name.into_inner())
-        .resolve_class_target(&pool)
-        .await?;
-    read_related_classes(pool, requestor, target, req).await
-}
-
-async fn read_related_classes(
-    pool: AppContext,
-    requestor: Authenticated,
-    target: ResolvedClassTarget,
-    req: HttpRequest,
-) -> Result<ApiResponse<Vec<HubuumClassWithPath>>, ApiError> {
-    let user = &requestor.principal;
-    let params = parse_query_parameter(req.query_string())?;
-    let class = target.class();
-    can!(
-        &pool,
-        user,
-        requestor.scopes(),
-        [Permissions::ReadClass],
-        class
-    );
-
-    let search_params = prepare_db_pagination::<ClassGraphRow>(&params)?;
-    let (classes, total_count) = user
-        .classes_related_to_page(&pool, class.clone(), search_params, requestor.scopes())
-        .await?;
-
-    ApiResponse::mapped_paginated(classes, total_count, &params, |page| {
-        page.to_descendant_classes_with_path()
-    })
-}
-
-// Contextual post for class relations
-#[utoipa::path(
-    post,
-    path = "/api/v1/classes/{class_id}/relations",
-    tag = "classes",
-    security(("bearer_auth" = [])),
-    params(
-        ("class_id" = i32, Path, description = "Class ID")
-    ),
-    request_body = NewHubuumClassRelationFromClass,
-    responses(
-        (status = 201, description = "Class relation created", body = HubuumClassRelation),
-        (status = 400, description = "Bad request", body = ApiErrorResponse),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 409, description = "Conflict", body = ApiErrorResponse)
-    )
-)]
-#[routes]
-#[post("/{class_id}/relations")]
-#[post("/{class_id}/relations/")]
-async fn create_class_relation(
-    pool: AppContext,
-    requestor: Authenticated,
-    class_id: web::Path<HubuumClassID>,
-    relation_data: web::Json<NewHubuumClassRelationFromClass>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    use crate::models::NewHubuumClassRelation;
-    use crate::traits::CollectionAccessors;
-
-    let user = &requestor.principal;
-    let class_id = class_id.into_inner();
-    let partial_relation = relation_data.into_inner();
-
-    debug!(
-        message = "Creating class relation",
-        user_id = user.id(),
-        from_class = class_id.id(),
-        to_class = partial_relation.to_hubuum_class_id,
-    );
-
-    let relation = NewHubuumClassRelation {
-        from_hubuum_class_id: class_id.id(),
-        to_hubuum_class_id: partial_relation.to_hubuum_class_id,
-        forward_template_alias: partial_relation.forward_template_alias.clone(),
-        reverse_template_alias: partial_relation.reverse_template_alias.clone(),
-    };
-
-    if pool.permission_backend().uses_sql_permission_store() {
-        let ids = relation.collection_id(&pool).await?;
-        can!(
-            &pool,
-            user,
-            requestor.scopes(),
-            [Permissions::CreateClassRelation],
-            ids.0,
-            ids.1
-        );
-    } else {
-        let resource = relation.to_resource_ref(&pool).await?;
-        authorize_resources(
-            pool.permission_backend(),
-            &pool,
-            user,
-            requestor.scopes(),
-            vec![Permissions::CreateClassRelation],
-            vec![resource],
-        )
-        .await?;
-    }
-
-    let event_context = requestor.event_context(&req);
-    let relation = relation.save(&pool, &event_context).await?;
-
-    let location = api_locations::class_relation(class_id.id(), relation.id())?;
-    Ok(ApiResponse::created(relation, location))
-}
-
-#[utoipa::path(
-    delete,
-    path = "/api/v1/classes/{class_id}/relations/{relation_id}",
-    tag = "classes",
-    security(("bearer_auth" = [])),
-    params(
-        ("class_id" = i32, Path, description = "Class ID"),
-        ("relation_id" = i32, Path, description = "Class relation ID")
-    ),
-    responses(
-        (status = 204, description = "Class relation deleted"),
-        (status = 400, description = "Bad request", body = ApiErrorResponse),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 404, description = "Class or relation not found", body = ApiErrorResponse)
-    )
-)]
-#[delete("/{class_id}/relations/{relation_id}")]
-async fn delete_class_relation(
-    pool: AppContext,
-    requestor: Authenticated,
-    paths: web::Path<(HubuumClassID, HubuumClassRelationID)>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    use crate::traits::CollectionAccessors;
-
-    let user = &requestor.principal;
-    let (class_id, relation_id) = paths.into_inner();
-
-    debug!(
-        message = "Deleting class relation",
-        user_id = user.id(),
-        class_id = class_id.id(),
-        relation_id = relation_id.id()
-    );
-
-    let relation = relation_id.instance(&pool).await?;
-
-    if pool.permission_backend().uses_sql_permission_store() {
-        let ids = relation_id.collection_id(&pool).await?;
-        can!(
-            &pool,
-            user,
-            requestor.scopes(),
-            [Permissions::DeleteClassRelation],
-            ids.0,
-            ids.1
-        );
-    } else {
-        let resource = relation.to_resource_ref(&pool).await?;
-        authorize_resources(
-            pool.permission_backend(),
-            &pool,
-            user,
-            requestor.scopes(),
-            vec![Permissions::DeleteClassRelation],
-            vec![resource],
-        )
-        .await?;
-    }
-
-    if relation.from_hubuum_class_id == class_id.id()
-        || relation.to_hubuum_class_id == class_id.id()
-    {
-        let event_context = requestor.event_context(&req);
-        relation.delete(&pool, &event_context).await?;
-        Ok(ApiResponse::no_content())
-    } else {
-        info!(
-            message = "Relation membership mismatch when deleting relation: class does not match either endpoint",
-            user_id = user.id(),
-            class_id = class_id.id(),
-            relation_id = relation_id.id(),
-            relation_from_class = relation.from_hubuum_class_id,
-            relation_to_class = relation.to_hubuum_class_id
-        );
-        Err(ApiError::BadRequest(format!(
-            "Class {} is not part of relation {}.",
-            class_id.id(),
-            relation.id,
-        )))
-    }
-}
-
-#[utoipa::path(
-    get,
-    path = "/api/v1/classes/{class_id}/related/relations",
-    tag = "classes",
-    security(("bearer_auth" = [])),
-    params(
-        ("class_id" = i32, Path, description = "Class ID")
-    ),
-    responses(
-        (status = 200, description = "Direct relations touching the class", body = [HubuumClassRelation]),
-        (status = 400, description = "Bad request", body = ApiErrorResponse),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 404, description = "Class not found", body = ApiErrorResponse)
-    )
-)]
-#[routes]
-#[get("/{class_id}/related/relations")]
-#[get("/{class_id}/related/relations/")]
-async fn get_related_class_relations(
-    pool: AppContext,
-    requestor: Authenticated,
-    class_id: web::Path<HubuumClassID>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    let target = ClassSelector::by_id(class_id.into_inner())
-        .resolve_class_target(&pool)
-        .await?;
-    read_related_class_relations(pool, requestor, target, req).await
-}
-
-#[utoipa::path(
-    get,
-    path = "/api/v1/classes/by-name/{class_name}/related/relations",
-    tag = "classes",
-    summary = "Get related class relations by class name",
-    security(("bearer_auth" = [])),
-    params(("class_name" = String, Path, description = "Globally unique class name")),
-    responses(
-        (status = 200, description = "Direct relations touching the class", body = [HubuumClassRelation]),
-        (status = 400, description = "Bad request", body = ApiErrorResponse),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 404, description = "Class not found", body = ApiErrorResponse)
-    )
-)]
-#[get("/by-name/{class_name}/related/relations")]
-async fn get_related_class_relations_by_name(
-    pool: AppContext,
-    requestor: Authenticated,
-    class_name: web::Path<String>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    let target = ClassSelector::by_name(class_name.into_inner())
-        .resolve_class_target(&pool)
-        .await?;
-    read_related_class_relations(pool, requestor, target, req).await
-}
-
-async fn read_related_class_relations(
-    pool: AppContext,
-    requestor: Authenticated,
-    target: ResolvedClassTarget,
-    req: HttpRequest,
-) -> Result<ApiResponse<Vec<HubuumClassRelation>>, ApiError> {
-    let user = &requestor.principal;
-    let params = parse_query_parameter(req.query_string())?;
-    let class = target.class();
-    can!(
-        &pool,
-        user,
-        requestor.scopes(),
-        [Permissions::ReadClass],
-        class
-    );
-
-    debug!(
-        message = "Getting direct relations touching class",
-        user_id = user.id(),
-        class_id = class.id
-    );
-
-    let (relations, total_count) = if pool.permission_backend().supports_sql_visibility_pushdown() {
-        let search_params = prepare_db_pagination::<HubuumClassRelation>(&params)?;
-        user.class_relations_touching_page(&pool, class.clone(), search_params, requestor.scopes())
-            .await?
-    } else {
-        let mut required = params.filters.permissions()?;
-        required.ensure_contains(&[Permissions::ReadClassRelation]);
-        let required = required.iter().copied().collect::<Vec<_>>();
-        if !scope_allows(requestor.scopes(), &required) {
-            return ApiResponse::paginated(Vec::new(), 0, &params);
-        }
-        let mut candidate_options = count_query_options(&params);
-        candidate_options.include_total = false;
-        let (candidates, _) = user
-            .class_relations_touching_page_from_backend_with_admin_status(
-                &pool,
-                class.clone(),
-                candidate_options,
-                true,
-                None,
-            )
-            .await?;
-        let resources = class_relation_authorization_resources(&pool, &candidates)
-            .await?
-            .into_iter()
-            .map(|resource| (resource.id, resource))
-            .collect::<HashMap<_, _>>();
-        let principal = PrincipalRef::load(&pool, user).await?;
-        let search_params = prepare_db_pagination::<HubuumClassRelation>(&params)?;
-        let page = authorize_cursor_page(
-            pool.permission_backend(),
-            &principal,
-            candidates,
-            required,
-            &search_params,
-            |relation| {
-                resources
-                    .get(&relation.id)
-                    .expect("every relation candidate has an authorization resource")
-                    .clone()
-            },
-        )
-        .await?;
-        (page.rows, page.total_count)
-    };
-    ApiResponse::paginated(relations, total_count, &params)
-}
-
-#[utoipa::path(
-    get,
-    path = "/api/v1/classes/{class_id}/related/graph",
-    tag = "classes",
-    security(("bearer_auth" = [])),
-    params(
-        ("class_id" = i32, Path, description = "Class ID")
-    ),
-    responses(
-        (status = 200, description = "Neighborhood graph for the class", body = RelatedClassGraph),
-        (status = 400, description = "Bad request", body = ApiErrorResponse),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 404, description = "Class not found", body = ApiErrorResponse)
-    )
-)]
-#[routes]
-#[get("/{class_id}/related/graph")]
-#[get("/{class_id}/related/graph/")]
-async fn get_related_class_graph(
-    pool: AppContext,
-    requestor: Authenticated,
-    class_id: web::Path<HubuumClassID>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    let target = ClassSelector::by_id(class_id.into_inner())
-        .resolve_class_target(&pool)
-        .await?;
-    read_related_class_graph(pool, requestor, target, req).await
-}
-
-#[utoipa::path(
-    get,
-    path = "/api/v1/classes/by-name/{class_name}/related/graph",
-    tag = "classes",
-    summary = "Get the related-class graph by class name",
-    security(("bearer_auth" = [])),
-    params(("class_name" = String, Path, description = "Globally unique class name")),
-    responses(
-        (status = 200, description = "Neighborhood graph for the class", body = RelatedClassGraph),
-        (status = 400, description = "Bad request", body = ApiErrorResponse),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 404, description = "Class not found", body = ApiErrorResponse)
-    )
-)]
-#[get("/by-name/{class_name}/related/graph")]
-async fn get_related_class_graph_by_name(
-    pool: AppContext,
-    requestor: Authenticated,
-    class_name: web::Path<String>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    let target = ClassSelector::by_name(class_name.into_inner())
-        .resolve_class_target(&pool)
-        .await?;
-    read_related_class_graph(pool, requestor, target, req).await
-}
-
-async fn read_related_class_graph(
-    pool: AppContext,
-    requestor: Authenticated,
-    target: ResolvedClassTarget,
-    req: HttpRequest,
-) -> Result<ApiResponse<RelatedClassGraph>, ApiError> {
-    let user = &requestor.principal;
-    let (params, graph_limit) =
-        prepare_graph_query_options(parse_query_parameter(req.query_string())?)?;
-    let class = target.class();
-    can!(
-        &pool,
-        user,
-        requestor.scopes(),
-        [Permissions::ReadClass],
-        class
-    );
-
-    let root_class = class_with_root_path(class);
-    let connected_classes = user
-        .search_classes_related_to(&pool, class.clone(), params, requestor.scopes())
-        .await?;
-    ensure_graph_result_within_limit(&connected_classes, graph_limit, "classes")?;
-    let mut classes = Vec::with_capacity(connected_classes.len() + 1);
-    classes.push(root_class);
-    classes.extend(connected_classes.to_descendant_classes_with_path());
-
-    let class_ids = classes.iter().map(|item| item.id).collect::<Vec<_>>();
-    let relations = user
-        .search_class_relations_between_ids(&pool, &class_ids, requestor.scopes())
-        .await?;
-
-    Ok(ApiResponse::new(
-        RelatedClassGraph { classes, relations },
-        StatusCode::OK,
-    ))
-}
-
-//
-// Object API
-//
-
-#[utoipa::path(
-    get,
-    path = "/api/v1/classes/{class_id}/",
-    tag = "classes",
-    description = "Lists objects in the path class. Enabled computed fields can be filtered with computed.shared.<key> or computed.personal.<key> using the normal __operator suffix, and sorted with the same names. Computed querying supports at most two computed filter parameters and two explicit sort fields per request.",
-    security(("bearer_auth" = [])),
-    params(
-        ("class_id" = i32, Path, description = "Class ID"),
-        ("include" = Option<String>, Query, description = "Set to computed to enrich each object"),
-        ("sort" = Option<String>, Query, description = "Sort by object fields or computed.shared.<key>/computed.personal.<key>; computed sorting supports at most two explicit sort fields")
-    ),
-    responses(
-        (status = 200, description = "Objects in class, optionally enriched with computed fields", body = [crate::models::HubuumObjectReadResponse]),
-        (status = 400, description = "Bad request", body = ApiErrorResponse),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 404, description = "Class not found", body = ApiErrorResponse)
-    )
-)]
-#[get("/{class_id}/")]
-async fn get_objects_in_class(
-    pool: AppContext,
-    requestor: Authenticated,
-    class_id: web::Path<HubuumClassID>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    let target = ClassSelector::by_id(class_id.into_inner())
-        .resolve_class_target(&pool)
-        .await?;
-    list_objects_in_resolved_class(pool, requestor, target, req).await
-}
-
-#[utoipa::path(
-    get,
-    path = "/api/v1/classes/by-name/{class_name}/objects",
-    tag = "classes",
-    summary = "List objects in a class by class name",
-    description = "Name-addressed alias for listing the current objects in a class. Numeric-looking class names remain names.",
-    security(("bearer_auth" = [])),
-    params(
-        ("class_name" = String, Path, description = "Globally unique class name"),
-        ("include" = Option<String>, Query, description = "Set to computed to enrich each object"),
-        ("sort" = Option<String>, Query, description = "Sort by object fields or computed.shared.<key>/computed.personal.<key>; computed sorting supports at most two explicit sort fields")
-    ),
-    responses(
-        (status = 200, description = "Objects in class, optionally enriched with computed fields", body = [crate::models::HubuumObjectReadResponse]),
-        (status = 400, description = "Bad request", body = ApiErrorResponse),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 404, description = "Class not found", body = ApiErrorResponse)
-    )
-)]
-#[get("/by-name/{class_name}/objects")]
-async fn get_objects_in_class_by_name(
-    pool: AppContext,
-    requestor: Authenticated,
-    class_name: web::Path<String>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    let target = ClassSelector::by_name(class_name.into_inner())
-        .resolve_class_target(&pool)
-        .await?;
-    list_objects_in_resolved_class(pool, requestor, target, req).await
-}
-
-async fn list_objects_in_resolved_class(
-    pool: AppContext,
-    requestor: Authenticated,
-    target: ResolvedClassTarget,
-    req: HttpRequest,
-) -> Result<ApiResponse<Vec<HubuumObjectReadResponse>>, ApiError> {
-    let user = &requestor.principal;
-    let class_id = HubuumClassID::new(target.class().id)?;
-    let class = target.class().clone();
-    let query_string = req.query_string();
-
-    let (mut params, include_computed) = parse_computed_object_list_query(query_string)?;
-
-    // The path is authoritative even if the caller supplied a conflicting
-    // class_id/classes filter.
-    scope_object_query_to_class(&mut params, &class_id);
-
-    debug!(
-        message = "Getting objects in class",
-        user_id = user.id(),
-        class_id = class_id.id(),
-        query = query_string
-    );
-
-    let computed_filtering = params
-        .filters
-        .iter()
-        .any(|filter| filter.field.computed_query().is_some());
-    let computed_sorting = params
-        .sort
-        .iter()
-        .any(|sort| sort.field.computed_query().is_some());
-    let computed_querying = computed_filtering || computed_sorting;
-    if computed_querying {
-        return computed_objects::list_objects(&pool, &requestor, &class, params, include_computed)
-            .await;
-    }
-
-    let (objects, total_count) = if pool.permission_backend().supports_sql_visibility_pushdown() {
-        let total_count = if params.include_total {
-            user.count_objects(&pool, count_query_options(&params), requestor.scopes())
-                .await?
-        } else {
-            SKIPPED_TOTAL_COUNT
-        };
-        let search_params = prepare_db_pagination::<HubuumObject>(&params)?;
-        let objects = user
-            .search_objects(&pool, search_params, requestor.scopes())
-            .await?;
-        (objects, total_count)
-    } else {
-        if !scope_allows(requestor.scopes(), &[Permissions::ReadObject]) {
-            let page = finalize_page(Vec::<HubuumObject>::new(), &params)?;
-            return object_read_page(page, 0, effective_page_limit(&params)?, include_computed);
-        }
-        let candidates = user
-            .search_objects_from_backend_with_admin_status(
-                &pool,
-                count_query_options(&params),
-                true,
-                None,
-            )
-            .await?;
-        let principal = PrincipalRef::load(&pool, user).await?;
-        let search_params = prepare_db_pagination::<HubuumObject>(&params)?;
-        let page = authorize_cursor_page(
-            pool.permission_backend(),
-            &principal,
-            candidates,
-            vec![Permissions::ReadObject],
-            &search_params,
-            |object| ResourceRef {
-                kind: ResourceKind::Object,
-                id: object.id,
-                attrs: ResourceAttrs {
-                    collection_id: Some(object.collection_id),
-                    class_id: Some(object.hubuum_class_id),
-                    name: Some(object.name.clone()),
-                    ..Default::default()
-                },
-            },
-        )
-        .await?;
-        (page.rows, page.total_count)
-    };
-
-    let page = finalize_page(objects, &params)?;
-    if include_computed {
-        let personal_owner = computed_personal_owner(&pool, &requestor, &class).await?;
-        let next_cursor = page.next_cursor;
-        let enriched = enrich_objects_with_computed(&pool, page.items, personal_owner).await?;
-        object_read_page(
-            Page {
-                items: enriched,
-                next_cursor,
-            },
-            total_count,
-            effective_page_limit(&params)?,
-            true,
-        )
-    } else {
-        object_read_page(page, total_count, effective_page_limit(&params)?, false)
-    }
-}
-
-#[utoipa::path(
-    post,
-    path = "/api/v1/classes/{class_id}/",
-    tag = "classes",
-    security(("bearer_auth" = [])),
-    params(
-        ("class_id" = i32, Path, description = "Class ID")
-    ),
-    request_body = NewHubuumObjectRequest,
-    responses(
-        (status = 201, description = "Object created in class", body = HubuumObject),
-        (status = 400, description = "Bad request", body = ApiErrorResponse),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 409, description = "Conflict", body = ApiErrorResponse)
-    )
-)]
-#[post("/{class_id}/")]
-async fn create_object_in_class(
-    pool: AppContext,
-    requestor: Authenticated,
-    class_id: web::Path<HubuumClassID>,
-    object_data: web::Json<NewHubuumObjectRequest>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    let target = ClassSelector::by_id(class_id.into_inner())
-        .resolve_class_target(&pool)
-        .await?;
-    let object =
-        create_object_in_resolved_class(&pool, &requestor, &req, target, object_data.into_inner())
-            .await?;
-    let location = api_locations::class_object(object.hubuum_class_id, object.id())?;
-    Ok(ApiResponse::created(object, location))
-}
-
-#[utoipa::path(
-    post,
-    path = "/api/v1/classes/by-name/{class_name}/objects",
-    tag = "classes",
-    summary = "Create an object in a class by class name",
-    description = "Name-addressed alias for object creation. The original class name and resolved ID are rechecked under the same transaction that creates the object.",
-    security(("bearer_auth" = [])),
-    params(("class_name" = String, Path, description = "Globally unique current class name")),
-    request_body = NewHubuumObjectRequest,
-    responses(
-        (status = 201, description = "Object created in class", body = HubuumObject),
-        (status = 400, description = "Bad request", body = ApiErrorResponse),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 403, description = "Forbidden", body = ApiErrorResponse),
-        (status = 404, description = "Class not found or concurrently renamed", body = ApiErrorResponse),
-        (status = 409, description = "Conflict", body = ApiErrorResponse)
-    )
-)]
-#[post("/by-name/{class_name}/objects")]
-async fn create_object_in_class_by_name(
-    pool: AppContext,
-    requestor: Authenticated,
-    class_name: web::Path<String>,
-    object_data: web::Json<NewHubuumObjectRequest>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    let target = ClassSelector::by_name(class_name.into_inner())
-        .resolve_class_target(&pool)
-        .await?;
-    let object =
-        create_object_in_resolved_class(&pool, &requestor, &req, target, object_data.into_inner())
-            .await?;
-    let location = api_locations::class_object(object.hubuum_class_id, object.id())?;
-    Ok(ApiResponse::created(object, location))
-}
-
-async fn create_object_in_resolved_class(
-    pool: &AppContext,
-    requestor: &Authenticated,
-    req: &HttpRequest,
-    target: ResolvedClassTarget,
-    object: NewHubuumObjectRequest,
-) -> Result<HubuumObject, ApiError> {
-    let object = object.into_domain_for_class(target.class())?;
-    debug!(
-        message = "Creating object in class",
-        user_id = requestor.principal.id(),
-        class_id = target.class().id,
-        object_name = object.name,
-    );
-
-    can!(
-        pool,
-        &requestor.principal,
-        requestor.scopes(),
-        [Permissions::CreateObject],
-        target.class()
-    );
-    let event_context = requestor.event_context(req);
-    object
-        .create_object_in_resolved_class(pool, &target, &event_context)
-        .await
-}
-
-async fn read_resolved_object(
-    pool: &AppContext,
-    requestor: &Authenticated,
-    req: &HttpRequest,
-    target: ResolvedObjectTarget,
-) -> Result<ApiResponse<HubuumObjectReadResponse>, ApiError> {
-    can!(
-        pool,
-        &requestor.principal,
-        requestor.scopes(),
-        [Permissions::ReadObject],
-        target.object()
-    );
-
-    let (_, include_computed) = parse_computed_include(req.query_string())?;
-    if include_computed {
-        let personal_owner = computed_personal_owner(pool, requestor, target.class()).await?;
-        let enriched =
-            enrich_objects_with_computed(pool, vec![target.object().clone()], personal_owner)
-                .await?
-                .pop()
-                .ok_or_else(|| {
-                    ApiError::InternalServerError(
-                        "Computed object enrichment returned no object".to_string(),
-                    )
-                })?;
-        Ok(ApiResponse::new_private_no_store(
-            HubuumObjectReadResponse::Computed(enriched),
-            StatusCode::OK,
-        ))
-    } else {
-        Ok(ApiResponse::new(
-            HubuumObjectReadResponse::Raw(target.object().clone()),
-            StatusCode::OK,
-        ))
-    }
-}
-
-#[utoipa::path(
-    get,
-    path = "/api/v1/classes/{class_id}/{object_id}",
-    tag = "classes",
-    security(("bearer_auth" = [])),
-    params(
-        ("class_id" = i32, Path, description = "Class ID"),
-        ("object_id" = i32, Path, description = "Object ID"),
-        ("include" = Option<String>, Query, description = "Set to computed to enrich the object")
-    ),
-    responses(
-        (status = 200, description = "Object in class, optionally enriched with computed fields", body = crate::models::HubuumObjectReadResponse),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 404, description = "Object not found", body = ApiErrorResponse)
-    )
-)]
-#[get("/{class_id}/{object_id}")]
-async fn get_object_in_class(
-    pool: AppContext,
-    requestor: Authenticated,
-    paths: web::Path<(HubuumClassID, HubuumObjectID)>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    let user = &requestor.principal;
-    let (class_id, object_id) = paths.into_inner();
-
-    debug!(
-        message = "Getting object in class",
-        user_id = user.id(),
-        class_id = class_id.id(),
-        object_id = object_id.id()
-    );
-
-    let target = ObjectSelector::by_id(class_id, object_id)
-        .resolve_object_target(&pool)
-        .await?;
-    read_resolved_object(&pool, &requestor, &req, target).await
-}
-
-#[utoipa::path(
-    get,
-    path = "/api/v1/classes/by-name/{class_name}/objects/by-name/{object_name}",
-    tag = "classes",
-    summary = "Get an object by class and object name",
-    description = "Fully name-addressed alias for the object endpoint. Numeric-looking path values remain names.",
-    security(("bearer_auth" = [])),
-    params(
-        ("class_name" = String, Path, description = "Globally unique class name"),
-        ("object_name" = String, Path, description = "Object name, unique within the class"),
-        ("include" = Option<String>, Query, description = "Set to computed to enrich the object")
-    ),
-    responses(
-        (status = 200, description = "Object in class, optionally enriched with computed fields", body = crate::models::HubuumObjectReadResponse),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 403, description = "Forbidden", body = ApiErrorResponse),
-        (status = 404, description = "Class/object name pair not found", body = ApiErrorResponse)
-    )
-)]
-#[get("/by-name/{class_name}/objects/by-name/{object_name}")]
-async fn get_object_in_class_by_name(
-    pool: AppContext,
-    requestor: Authenticated,
-    paths: web::Path<(String, String)>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    let (class_name, object_name) = paths.into_inner();
-    let target = ObjectSelector::by_name(class_name, object_name)
-        .resolve_object_target(&pool)
-        .await?;
-    read_resolved_object(&pool, &requestor, &req, target).await
-}
-
-async fn apply_resolved_object_update(
-    pool: &AppContext,
-    requestor: &Authenticated,
-    req: &HttpRequest,
-    target: ResolvedObjectTarget,
-    update: UpdateHubuumObject,
-) -> Result<HubuumObject, ApiError> {
-    can!(
-        pool,
-        &requestor.principal,
-        requestor.scopes(),
-        [Permissions::UpdateObject],
-        target.object()
-    );
-    ensure_object_update_stays_in_path_class(&update, target.object())?;
-
-    let event_context = requestor.event_context(req);
-    update
-        .update_resolved_object(pool, &target, &event_context)
-        .await
-}
-
-#[utoipa::path(
-    patch,
-    path = "/api/v1/classes/{class_id}/{object_id}",
-    tag = "classes",
-    security(("bearer_auth" = [])),
-    params(
-        ("class_id" = i32, Path, description = "Class ID"),
-        ("object_id" = i32, Path, description = "Object ID")
-    ),
-    request_body = UpdateHubuumObjectRequest,
-    responses(
-        (status = 200, description = "Updated object", body = HubuumObject),
-        (status = 400, description = "Bad request", body = ApiErrorResponse),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 404, description = "Object not found", body = ApiErrorResponse)
-    )
-)]
-#[patch("/{class_id}/{object_id}")]
-async fn patch_object_in_class(
-    pool: AppContext,
-    requestor: Authenticated,
-    paths: web::Path<(HubuumClassID, HubuumObjectID)>,
-    object_data: web::Json<UpdateHubuumObjectRequest>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    let user = &requestor.principal;
-    let (class_id, object_id) = paths.into_inner();
-    let object_data = object_data.into_inner().into_domain()?;
-
-    debug!(
-        message = "Updating object in class",
-        user_id = user.id(),
-        class_id = class_id.id(),
-        object_id = object_id.id()
-    );
-
-    let target = ObjectSelector::by_id(class_id, object_id)
-        .resolve_object_target(&pool)
-        .await?;
-    let object = apply_resolved_object_update(&pool, &requestor, &req, target, object_data).await?;
-    Ok(ApiResponse::new(object, StatusCode::OK))
-}
-
-#[utoipa::path(
-    patch,
-    path = "/api/v1/classes/by-name/{class_name}/objects/by-name/{object_name}",
-    tag = "classes",
-    summary = "Update an object by class and object name",
-    description = "Fully name-addressed alias for object update. The original natural key and resolved IDs are rechecked under the transactional row lock.",
-    security(("bearer_auth" = [])),
-    params(
-        ("class_name" = String, Path, description = "Globally unique current class name"),
-        ("object_name" = String, Path, description = "Current object name, unique within the class")
-    ),
-    request_body = UpdateHubuumObjectRequest,
-    responses(
-        (status = 200, description = "Updated object", body = HubuumObject),
-        (status = 400, description = "Bad request", body = ApiErrorResponse),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 403, description = "Forbidden", body = ApiErrorResponse),
-        (status = 404, description = "Class/object pair not found or concurrently renamed", body = ApiErrorResponse)
-    )
-)]
-#[patch("/by-name/{class_name}/objects/by-name/{object_name}")]
-async fn patch_object_in_class_by_name(
-    pool: AppContext,
-    requestor: Authenticated,
-    paths: web::Path<(String, String)>,
-    object_data: web::Json<UpdateHubuumObjectRequest>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    let (class_name, object_name) = paths.into_inner();
-    let target = ObjectSelector::by_name(class_name, object_name)
-        .resolve_object_target(&pool)
-        .await?;
-    let object = apply_resolved_object_update(
-        &pool,
-        &requestor,
-        &req,
-        target,
-        object_data.into_inner().into_domain()?,
-    )
-    .await?;
-    Ok(ApiResponse::new(object, StatusCode::OK))
-}
-
-async fn apply_object_data_patch(
-    pool: &AppContext,
-    requestor: &Authenticated,
-    req: &HttpRequest,
-    target: ResolvedObjectTarget,
-    patch: ObjectDataPatchDocument,
-) -> Result<HubuumObject, ApiError> {
-    can!(
-        pool,
-        &requestor.principal,
-        requestor.scopes(),
-        [Permissions::UpdateObject],
-        target.object()
-    );
-
-    let event_context = requestor.event_context(req);
-    patch.patch_object_data(pool, &target, &event_context).await
-}
-
-#[utoipa::path(
-    patch,
-    path = "/api/v1/classes/{class_id}/{object_id}/data",
-    tag = "classes",
-    summary = "Apply JSON Patch to object data",
-    description = "Atomically applies an RFC 6902 JSON Patch document to the object's raw `data` document. Every `path` and `from` is an RFC 6901 JSON Pointer relative to the root of `data`; an empty path replaces the complete document. Operations are evaluated against the latest row-locked data. A successful change updates schema validation, computed-field materialization, history, timestamps, and the object update event in one transaction. A patch that produces no data change returns the current object without changing its timestamp or emitting an update event.",
-    security(("bearer_auth" = [])),
-    params(
-        ("class_id" = i32, Path, description = "Class ID"),
-        ("object_id" = i32, Path, description = "Object ID")
-    ),
-    request_body(
-        content = ObjectDataPatchDocument,
-        content_type = "application/json-patch+json",
-        description = "RFC 6902 operation array. Supports add, remove, replace, move, copy, and test. Limited to 1,000 operations and 128 pointer segments per path/from."
-    ),
-    responses(
-        (status = 200, description = "Updated object, or the unchanged object for a no-op patch", body = HubuumObject),
-        (status = 400, description = "Malformed JSON or invalid JSON Patch structure, operation count, or pointer depth", body = ApiErrorResponse),
-        (status = 401, description = "Missing or invalid authentication", body = ApiErrorResponse),
-        (status = 403, description = "UpdateObject permission denied", body = ApiErrorResponse),
-        (status = 404, description = "Class/object pair not found", body = ApiErrorResponse),
-        (status = 406, description = "Final patched data fails the class JSON Schema", body = ApiErrorResponse),
-        (status = 409, description = "A patch operation failed, including a failed test operation; no change was persisted", body = ApiErrorResponse),
-        (status = 413, description = "JSON Patch request or resulting object data exceeds its resource limits", body = ApiErrorResponse),
-        (status = 415, description = "Content-Type is not application/json-patch+json", body = ApiErrorResponse),
-        (status = 500, description = "Persistence, computed-field materialization, or event emission failed and the transaction was rolled back", body = ApiErrorResponse)
-    )
-)]
-#[patch("/{class_id}/{object_id}/data")]
-async fn patch_object_data_in_class(
-    pool: AppContext,
-    requestor: Authenticated,
-    paths: web::Path<(HubuumClassID, HubuumObjectID)>,
-    patch: ObjectDataPatchPayload,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    let user = &requestor.principal;
-    let (class_id, object_id) = paths.into_inner();
-
-    debug!(
-        message = "Applying JSON Patch to object data",
-        user_id = user.id(),
-        class_id = class_id.id(),
-        object_id = object_id.id()
-    );
-
-    let target = ObjectSelector::by_id(class_id, object_id)
-        .resolve_object_target(&pool)
-        .await?;
-    let object =
-        apply_object_data_patch(&pool, &requestor, &req, target, patch.into_inner()).await?;
-    Ok(ApiResponse::new(object, StatusCode::OK))
-}
-
-#[utoipa::path(
-    patch,
-    path = "/api/v1/classes/by-name/{class_name}/objects/by-name/{object_name}/data",
-    tag = "classes",
-    summary = "Apply JSON Patch to object data by object name",
-    description = "Fully name-addressed alias for the atomic object-data JSON Patch endpoint. Class names are globally unique, object names are unique within their class, and both names are URL percent-decoded. The server resolves the natural key for authorization, then requires that same object ID, class ID, class name, and object name when taking the transactional row lock so a concurrent rename cannot redirect the patch. Numeric-looking names are always treated as names on this route. Every `path` and `from` remains relative to the root of the raw `data` document.",
-    security(("bearer_auth" = [])),
-    params(
-        ("class_name" = String, Path, description = "Globally unique class name, URL percent-encoded as needed"),
-        ("object_name" = String, Path, description = "Class-scoped object name, URL percent-encoded as needed")
-    ),
-    request_body(
-        content = ObjectDataPatchDocument,
-        content_type = "application/json-patch+json",
-        description = "RFC 6902 operation array. Supports add, remove, replace, move, copy, and test. Limited to 1,000 operations and 128 pointer segments per path/from."
-    ),
-    responses(
-        (status = 200, description = "Updated object, or the unchanged object for a no-op patch", body = HubuumObject),
-        (status = 400, description = "Malformed JSON or invalid JSON Patch structure, operation count, or pointer depth", body = ApiErrorResponse),
-        (status = 401, description = "Missing or invalid authentication", body = ApiErrorResponse),
-        (status = 403, description = "UpdateObject permission denied", body = ApiErrorResponse),
-        (status = 404, description = "Class-name/object-name pair not found or the object was concurrently renamed", body = ApiErrorResponse),
-        (status = 406, description = "Final patched data fails the class JSON Schema", body = ApiErrorResponse),
-        (status = 409, description = "A patch operation failed, including a failed test operation; no change was persisted", body = ApiErrorResponse),
-        (status = 413, description = "JSON Patch request or resulting object data exceeds its resource limits", body = ApiErrorResponse),
-        (status = 415, description = "Content-Type is not application/json-patch+json", body = ApiErrorResponse),
-        (status = 500, description = "Persistence, computed-field materialization, or event emission failed and the transaction was rolled back", body = ApiErrorResponse)
-    )
-)]
-#[patch("/by-name/{class_name}/objects/by-name/{object_name}/data")]
-async fn patch_object_data_by_name_in_class(
-    pool: AppContext,
-    requestor: Authenticated,
-    paths: web::Path<(String, String)>,
-    patch: ObjectDataPatchPayload,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    let user = &requestor.principal;
-    let (class_name, object_name) = paths.into_inner();
-    let target = ObjectSelector::by_name(class_name, object_name)
-        .resolve_object_target(&pool)
-        .await?;
-
-    debug!(
-        message = "Applying JSON Patch to object data by name",
-        user_id = user.id(),
-        class_id = target.object().hubuum_class_id,
-        object_id = target.object().id
-    );
-
-    let object =
-        apply_object_data_patch(&pool, &requestor, &req, target, patch.into_inner()).await?;
-    Ok(ApiResponse::new(object, StatusCode::OK))
-}
-
-async fn delete_resolved_object(
-    pool: &AppContext,
-    requestor: &Authenticated,
-    req: &HttpRequest,
-    target: ResolvedObjectTarget,
-) -> Result<(), ApiError> {
-    can!(
-        pool,
-        &requestor.principal,
-        requestor.scopes(),
-        [Permissions::DeleteObject],
-        target.object()
-    );
-
-    let event_context = requestor.event_context(req);
-    target.delete_resolved_object(pool, &event_context).await
-}
-
-#[utoipa::path(
-    delete,
-    path = "/api/v1/classes/{class_id}/{object_id}",
-    tag = "classes",
-    security(("bearer_auth" = [])),
-    params(
-        ("class_id" = i32, Path, description = "Class ID"),
-        ("object_id" = i32, Path, description = "Object ID")
-    ),
-    responses(
-        (status = 204, description = "Object deleted"),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 404, description = "Object not found", body = ApiErrorResponse)
-    )
-)]
-#[delete("/{class_id}/{object_id}")]
-async fn delete_object_in_class(
-    pool: AppContext,
-    requestor: Authenticated,
-    paths: web::Path<(HubuumClassID, HubuumObjectID)>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    let user = &requestor.principal;
-    let (class_id, object_id) = paths.into_inner();
-
-    debug!(
-        message = "Deleting object in class",
-        user_id = user.id(),
-        class_id = class_id.id(),
-        object_id = object_id.id()
-    );
-
-    let target = ObjectSelector::by_id(class_id, object_id)
-        .resolve_object_target(&pool)
-        .await?;
-    delete_resolved_object(&pool, &requestor, &req, target).await?;
-    Ok(ApiResponse::no_content())
-}
-
-#[utoipa::path(
-    delete,
-    path = "/api/v1/classes/by-name/{class_name}/objects/by-name/{object_name}",
-    tag = "classes",
-    summary = "Delete an object by class and object name",
-    description = "Fully name-addressed alias for object deletion. The original natural key and resolved IDs are rechecked under the transactional row lock.",
-    security(("bearer_auth" = [])),
-    params(
-        ("class_name" = String, Path, description = "Globally unique current class name"),
-        ("object_name" = String, Path, description = "Current object name, unique within the class")
-    ),
-    responses(
-        (status = 204, description = "Object deleted"),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 403, description = "Forbidden", body = ApiErrorResponse),
-        (status = 404, description = "Class/object pair not found or concurrently renamed", body = ApiErrorResponse)
-    )
-)]
-#[delete("/by-name/{class_name}/objects/by-name/{object_name}")]
-async fn delete_object_in_class_by_name(
-    pool: AppContext,
-    requestor: Authenticated,
-    paths: web::Path<(String, String)>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    let (class_name, object_name) = paths.into_inner();
-    let target = ObjectSelector::by_name(class_name, object_name)
-        .resolve_object_target(&pool)
-        .await?;
-    delete_resolved_object(&pool, &requestor, &req, target).await?;
-    Ok(ApiResponse::no_content())
-}
-
-#[utoipa::path(
-    get,
-    path = "/api/v1/classes/{class_id}/objects/{object_id}/related/objects",
-    tag = "classes",
-    security(("bearer_auth" = [])),
-    params(
-        ("class_id" = i32, Path, description = "Class ID"),
-        ("object_id" = i32, Path, description = "Object ID"),
-        ("ignore_classes" = Option<String>, Query, description = "Comma-separated class IDs to exclude from the returned connected objects"),
-        ("ignore_self_class" = Option<bool>, Query, description = "Exclude connected objects in the same class as the root object. Defaults to true")
-    ),
-    responses(
-        (status = 200, description = "Objects connected to the object", body = [HubuumObjectWithPath]),
-        (status = 400, description = "Bad request", body = ApiErrorResponse),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 404, description = "Class or object not found", body = ApiErrorResponse)
-    )
-)]
-#[routes]
-#[get("/{class_id}/objects/{object_id}/related/objects")]
-#[get("/{class_id}/objects/{object_id}/related/objects/")]
-async fn get_related_objects(
-    pool: AppContext,
-    requestor: Authenticated,
-    paths: web::Path<(HubuumClassID, HubuumObjectID)>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    let (class_id, object_id) = paths.into_inner();
-    let target = ObjectSelector::by_id(class_id, object_id)
-        .resolve_object_target(&pool)
-        .await?;
-    read_related_objects(pool, requestor, target, req).await
-}
-
-#[utoipa::path(
-    get,
-    path = "/api/v1/classes/by-name/{class_name}/objects/by-name/{object_name}/related/objects",
-    tag = "classes",
-    summary = "Get related objects by class and object name",
-    security(("bearer_auth" = [])),
-    params(
-        ("class_name" = String, Path, description = "Globally unique class name"),
-        ("object_name" = String, Path, description = "Object name, unique within the class"),
-        ("ignore_classes" = Option<String>, Query, description = "Comma-separated class IDs to exclude from the returned connected objects"),
-        ("ignore_self_class" = Option<bool>, Query, description = "Exclude connected objects in the same class as the root object. Defaults to true")
-    ),
-    responses(
-        (status = 200, description = "Objects connected to the object", body = [HubuumObjectWithPath]),
-        (status = 400, description = "Bad request", body = ApiErrorResponse),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 404, description = "Class/object name pair not found", body = ApiErrorResponse)
-    )
-)]
-#[get("/by-name/{class_name}/objects/by-name/{object_name}/related/objects")]
-async fn get_related_objects_by_name(
-    pool: AppContext,
-    requestor: Authenticated,
-    paths: web::Path<(String, String)>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    let (class_name, object_name) = paths.into_inner();
-    let target = ObjectSelector::by_name(class_name, object_name)
-        .resolve_object_target(&pool)
-        .await?;
-    read_related_objects(pool, requestor, target, req).await
-}
-
-async fn read_related_objects(
-    pool: AppContext,
-    requestor: Authenticated,
-    target: ResolvedObjectTarget,
-    req: HttpRequest,
-) -> Result<ApiResponse<Vec<HubuumObjectWithPath>>, ApiError> {
-    let user = &requestor.principal;
-    let object = target.object();
-    let query_string = req.query_string();
-
-    let (mut params, related_options) = parse_related_objects_query(query_string)?;
-
-    can!(
-        &pool,
-        user,
-        requestor.scopes(),
-        [Permissions::ReadObject],
-        object
-    );
-
-    if related_options.ignore_self_class {
-        params.filters.add_filter(
-            FilterField::ClassId,
-            SearchOperator::Equals { is_negated: true },
-            &object.hubuum_class_id.to_string(),
-        );
-    }
-
-    if !related_options.ignore_classes.is_empty() {
-        params.filters.add_filter(
-            FilterField::ClassId,
-            SearchOperator::Equals { is_negated: true },
-            &related_options
-                .ignore_classes
-                .iter()
-                .map(i32::to_string)
-                .collect::<Vec<_>>()
-                .join(","),
-        );
-    }
-
-    debug!(
-        message = "Getting objects connected to object",
-        user_id = user.id(),
-        class_id = object.hubuum_class_id,
-        object_id = object.id(),
-        query = query_string,
-        ignore_classes = ?related_options.ignore_classes,
-        ignore_self_class = related_options.ignore_self_class,
-    );
-
-    let search_params = prepare_db_pagination::<RelatedObjectGraphRow>(&params)?;
-    let (hits, total_count) = user
-        .objects_related_to_page(&pool, object.clone(), search_params, requestor.scopes())
-        .await?;
-
-    ApiResponse::mapped_paginated(hits, total_count, &params, |page| {
-        page.to_descendant_objects_with_path()
-    })
-}
-
-#[utoipa::path(
-    get,
-    path = "/api/v1/classes/{class_id}/objects/{object_id}/related/relations",
-    tag = "classes",
-    security(("bearer_auth" = [])),
-    params(
-        ("class_id" = i32, Path, description = "Class ID"),
-        ("object_id" = i32, Path, description = "Object ID")
-    ),
-    responses(
-        (status = 200, description = "Direct relations touching the object", body = [HubuumObjectRelation]),
-        (status = 400, description = "Bad request", body = ApiErrorResponse),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 404, description = "Class or object not found", body = ApiErrorResponse)
-    )
-)]
-#[routes]
-#[get("/{class_id}/objects/{object_id}/related/relations")]
-#[get("/{class_id}/objects/{object_id}/related/relations/")]
-async fn get_related_object_relations(
-    pool: AppContext,
-    requestor: Authenticated,
-    paths: web::Path<(HubuumClassID, HubuumObjectID)>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    let (class_id, object_id) = paths.into_inner();
-    let target = ObjectSelector::by_id(class_id, object_id)
-        .resolve_object_target(&pool)
-        .await?;
-    read_related_object_relations(pool, requestor, target, req).await
-}
-
-#[utoipa::path(
-    get,
-    path = "/api/v1/classes/by-name/{class_name}/objects/by-name/{object_name}/related/relations",
-    tag = "classes",
-    summary = "Get related object relations by class and object name",
-    security(("bearer_auth" = [])),
-    params(
-        ("class_name" = String, Path, description = "Globally unique class name"),
-        ("object_name" = String, Path, description = "Object name, unique within the class")
-    ),
-    responses(
-        (status = 200, description = "Direct relations touching the object", body = [HubuumObjectRelation]),
-        (status = 400, description = "Bad request", body = ApiErrorResponse),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 404, description = "Class/object name pair not found", body = ApiErrorResponse)
-    )
-)]
-#[get("/by-name/{class_name}/objects/by-name/{object_name}/related/relations")]
-async fn get_related_object_relations_by_name(
-    pool: AppContext,
-    requestor: Authenticated,
-    paths: web::Path<(String, String)>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    let (class_name, object_name) = paths.into_inner();
-    let target = ObjectSelector::by_name(class_name, object_name)
-        .resolve_object_target(&pool)
-        .await?;
-    read_related_object_relations(pool, requestor, target, req).await
-}
-
-async fn read_related_object_relations(
-    pool: AppContext,
-    requestor: Authenticated,
-    target: ResolvedObjectTarget,
-    req: HttpRequest,
-) -> Result<ApiResponse<Vec<HubuumObjectRelation>>, ApiError> {
-    let user = &requestor.principal;
-    let params = parse_query_parameter(req.query_string())?;
-    let object = target.object();
-    can!(
-        &pool,
-        user,
-        requestor.scopes(),
-        [Permissions::ReadObject],
-        object
-    );
-
-    debug!(
-        message = "Getting direct relations touching object",
-        user_id = user.id(),
-        class_id = object.hubuum_class_id,
-        object_id = object.id(),
-        query = req.query_string(),
-    );
-
-    let (relations, total_count) = if pool.permission_backend().supports_sql_visibility_pushdown() {
-        let search_params = prepare_db_pagination::<HubuumObjectRelation>(&params)?;
-        user.object_relations_touching_page(
-            &pool,
-            object.clone(),
-            search_params,
-            requestor.scopes(),
-        )
-        .await?
-    } else {
-        let mut required = params.filters.permissions()?;
-        required.ensure_contains(&[Permissions::ReadObjectRelation]);
-        let required = required.iter().copied().collect::<Vec<_>>();
-        if !scope_allows(requestor.scopes(), &required) {
-            return ApiResponse::paginated(Vec::new(), 0, &params);
-        }
-        let mut candidate_options = count_query_options(&params);
-        candidate_options.include_total = false;
-        let (candidates, _) = user
-            .object_relations_touching_page_from_backend_with_admin_status(
-                &pool,
-                object.clone(),
-                candidate_options,
-                true,
-                None,
-            )
-            .await?;
-        let resources = object_relation_authorization_resources(&pool, &candidates)
-            .await?
-            .into_iter()
-            .map(|resource| (resource.id, resource))
-            .collect::<HashMap<_, _>>();
-        let principal = PrincipalRef::load(&pool, user).await?;
-        let search_params = prepare_db_pagination::<HubuumObjectRelation>(&params)?;
-        let page = authorize_cursor_page(
-            pool.permission_backend(),
-            &principal,
-            candidates,
-            required,
-            &search_params,
-            |relation| {
-                resources
-                    .get(&relation.id)
-                    .expect("every relation candidate has an authorization resource")
-                    .clone()
-            },
-        )
-        .await?;
-        (page.rows, page.total_count)
-    };
-
-    ApiResponse::paginated(relations, total_count, &params)
-}
-
-#[utoipa::path(
-    get,
-    path = "/api/v1/classes/{class_id}/objects/{object_id}/related/graph",
-    tag = "classes",
-    security(("bearer_auth" = [])),
-    params(
-        ("class_id" = i32, Path, description = "Class ID"),
-        ("object_id" = i32, Path, description = "Object ID")
-    ),
-    responses(
-        (status = 200, description = "Neighborhood graph for the object", body = RelatedObjectGraph),
-        (status = 400, description = "Bad request", body = ApiErrorResponse),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 404, description = "Class or object not found", body = ApiErrorResponse)
-    )
-)]
-#[routes]
-#[get("/{class_id}/objects/{object_id}/related/graph")]
-#[get("/{class_id}/objects/{object_id}/related/graph/")]
-async fn get_related_object_graph(
-    pool: AppContext,
-    requestor: Authenticated,
-    paths: web::Path<(HubuumClassID, HubuumObjectID)>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    let (class_id, object_id) = paths.into_inner();
-    let target = ObjectSelector::by_id(class_id, object_id)
-        .resolve_object_target(&pool)
-        .await?;
-    read_related_object_graph(pool, requestor, target, req).await
-}
-
-#[utoipa::path(
-    get,
-    path = "/api/v1/classes/by-name/{class_name}/objects/by-name/{object_name}/related/graph",
-    tag = "classes",
-    summary = "Get the related-object graph by class and object name",
-    security(("bearer_auth" = [])),
-    params(
-        ("class_name" = String, Path, description = "Globally unique class name"),
-        ("object_name" = String, Path, description = "Object name, unique within the class")
-    ),
-    responses(
-        (status = 200, description = "Neighborhood graph for the object", body = RelatedObjectGraph),
-        (status = 400, description = "Bad request", body = ApiErrorResponse),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 404, description = "Class/object name pair not found", body = ApiErrorResponse)
-    )
-)]
-#[get("/by-name/{class_name}/objects/by-name/{object_name}/related/graph")]
-async fn get_related_object_graph_by_name(
-    pool: AppContext,
-    requestor: Authenticated,
-    paths: web::Path<(String, String)>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    let (class_name, object_name) = paths.into_inner();
-    let target = ObjectSelector::by_name(class_name, object_name)
-        .resolve_object_target(&pool)
-        .await?;
-    read_related_object_graph(pool, requestor, target, req).await
-}
-
-async fn read_related_object_graph(
-    pool: AppContext,
-    requestor: Authenticated,
-    target: ResolvedObjectTarget,
-    req: HttpRequest,
-) -> Result<ApiResponse<RelatedObjectGraph>, ApiError> {
-    let user = &requestor.principal;
-    let object = target.object();
-    let (params, graph_limit) =
-        prepare_graph_query_options(parse_query_parameter(req.query_string())?)?;
-
-    can!(
-        &pool,
-        user,
-        requestor.scopes(),
-        [Permissions::ReadObject],
-        object
-    );
-
-    debug!(
-        message = "Getting related object graph",
-        user_id = user.id(),
-        class_id = object.hubuum_class_id,
-        object_id = object.id(),
-        query = req.query_string(),
-    );
-
-    let root_object = object_with_root_path(object);
-    let connected_objects = user
-        .search_objects_related_to(&pool, object.clone(), params, requestor.scopes())
-        .await?;
-    ensure_graph_result_within_limit(&connected_objects, graph_limit, "objects")?;
-    let mut objects = Vec::with_capacity(connected_objects.len() + 1);
-    objects.push(root_object);
-    objects.extend(connected_objects.to_descendant_objects_with_path());
-
-    let object_ids = objects.iter().map(|item| item.id).collect::<Vec<_>>();
-    let relations = user
-        .search_object_relations_between_ids(&pool, &object_ids, requestor.scopes())
-        .await?;
-
-    Ok(ApiResponse::new(
-        RelatedObjectGraph { objects, relations },
-        StatusCode::OK,
-    ))
-}
-
-#[utoipa::path(
-    get,
-    path = "/api/v1/classes/{class_id}/{from_object_id}/relations/{to_class_id}/{to_object_id}",
-    tag = "classes",
-    security(("bearer_auth" = [])),
-    params(
-        ("class_id" = i32, Path, description = "Source class ID"),
-        ("from_object_id" = i32, Path, description = "Source object ID"),
-        ("to_class_id" = i32, Path, description = "Target class ID"),
-        ("to_object_id" = i32, Path, description = "Target object ID")
-    ),
-    responses(
-        (status = 200, description = "Object relation", body = HubuumObjectRelation),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 404, description = "Relation not found", body = ApiErrorResponse)
-    )
-)]
-#[get("/{class_id}/{from_object_id}/relations/{to_class_id}/{to_object_id}")]
-async fn get_object_relation_from_class_and_objects(
-    pool: AppContext,
-    requestor: Authenticated,
-    paths: web::Path<(HubuumClassID, HubuumObjectID, HubuumClassID, HubuumObjectID)>,
-) -> Result<impl Responder, ApiError> {
-    let user = &requestor.principal;
-    let (from_class, from_object, to_class, to_object) = paths.into_inner();
-
-    debug!(
-        message = "Getting object relation from class and objects",
-        user_id = user.id(),
-        class_id = from_class.id(),
-        from_object_id = from_object.id(),
-        to_object_id = to_object.id()
-    );
-
-    check_if_object_in_class(&pool, &from_class, &from_object).await?;
-    check_if_object_in_class(&pool, &to_class, &to_object).await?;
-
-    let relation = from_object
-        .object_relation(&pool, &from_class, &to_object)
-        .await
-        .map_err(|_| {
-            ApiError::NotFound(format!(
-                "Object {} of class {} is not related to object {}",
-                from_object.id(),
-                from_class.id(),
-                to_object.id()
-            ))
-        })?;
-    if pool.permission_backend().uses_sql_permission_store() {
-        can!(
-            &pool,
-            user,
-            requestor.scopes(),
-            [Permissions::ReadObjectRelation],
-            from_class,
-            from_object,
-            to_class,
-            to_object
-        );
-    } else {
-        let resource = relation.to_resource_ref(&pool).await?;
-        authorize_resources(
-            pool.permission_backend(),
-            &pool,
-            user,
-            requestor.scopes(),
-            vec![Permissions::ReadObjectRelation],
-            vec![resource],
-        )
-        .await?;
-    }
-    Ok(ApiResponse::new(relation, StatusCode::OK))
-}
-
-#[utoipa::path(
-    delete,
-    path = "/api/v1/classes/{class_id}/{from_object_id}/relations/{to_class_id}/{to_object_id}",
-    tag = "classes",
-    security(("bearer_auth" = [])),
-    params(
-        ("class_id" = i32, Path, description = "Source class ID"),
-        ("from_object_id" = i32, Path, description = "Source object ID"),
-        ("to_class_id" = i32, Path, description = "Target class ID"),
-        ("to_object_id" = i32, Path, description = "Target object ID")
-    ),
-    responses(
-        (status = 204, description = "Object relation deleted"),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 404, description = "Relation not found", body = ApiErrorResponse)
-    )
-)]
-#[delete("/{class_id}/{from_object_id}/relations/{to_class_id}/{to_object_id}")]
-async fn delete_object_relation(
-    pool: AppContext,
-    requestor: Authenticated,
-    paths: web::Path<(HubuumClassID, HubuumObjectID, HubuumClassID, HubuumObjectID)>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    let user = &requestor.principal;
-    let (from_class, from_object, to_class, to_object) = paths.into_inner();
-
-    check_if_object_in_class(&pool, &from_class, &from_object).await?;
-    check_if_object_in_class(&pool, &to_class, &to_object).await?;
-
-    debug!(
-        message = "Deleting object relation",
-        user_id = user.id(),
-        from_class_id = from_class.id(),
-        from_object_id = from_object.id(),
-        to_class_id = to_class.id(),
-        to_object_id = to_object.id()
-    );
-
-    let relation = from_object
-        .object_relation(&pool, &from_class, &to_object)
-        .await;
-
-    if relation.is_err() {
-        debug!(
-            message = "Relation does not exist",
-            user_id = user.id(),
-            from_class_id = from_class.id(),
-            from_object_id = from_object.id(),
-            to_class_id = to_class.id(),
-            to_object_id = to_object.id()
-        );
-        return Err(ApiError::NotFound(format!(
-            "Class {} is not related to class {}",
-            from_class.id(),
-            to_class.id()
-        )));
-    }
-
-    let relation = relation.expect("Relation should exist after is_err check");
-
-    if pool.permission_backend().uses_sql_permission_store() {
-        can!(
-            &pool,
-            user,
-            requestor.scopes(),
-            [Permissions::DeleteObjectRelation],
-            from_class,
-            from_object,
-            to_class,
-            to_object
-        );
-    } else {
-        let resource = relation.to_resource_ref(&pool).await?;
-        authorize_resources(
-            pool.permission_backend(),
-            &pool,
-            user,
-            requestor.scopes(),
-            vec![Permissions::DeleteObjectRelation],
-            vec![resource],
-        )
-        .await?;
-    }
-
-    debug!(
-        message = "Relation ID found",
-        user_id = user.id(),
-        class_id = from_class.id(),
-        object_id = from_object.id(),
-        relation_id = relation.id(),
-        relation_id_actual = relation.id()
-    );
-
-    let event_context = requestor.event_context(&req);
-    relation.delete(&pool, &event_context).await?;
-    Ok(ApiResponse::no_content())
-}
-
-#[utoipa::path(
-    post,
-    path = "/api/v1/classes/{class_id}/{from_object_id}/relations/{to_class_id}/{to_object_id}",
-    tag = "classes",
-    security(("bearer_auth" = [])),
-    params(
-        ("class_id" = i32, Path, description = "Source class ID"),
-        ("from_object_id" = i32, Path, description = "Source object ID"),
-        ("to_class_id" = i32, Path, description = "Target class ID"),
-        ("to_object_id" = i32, Path, description = "Target object ID")
-    ),
-    responses(
-        (status = 201, description = "Object relation created", body = HubuumObjectRelation),
-        (status = 400, description = "Bad request", body = ApiErrorResponse),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 404, description = "Class or object not found", body = ApiErrorResponse),
-        (status = 409, description = "Conflict", body = ApiErrorResponse)
-    )
-)]
-#[post("/{class_id}/{from_object_id}/relations/{to_class_id}/{to_object_id}")]
-async fn create_object_relation(
-    pool: AppContext,
-    requestor: Authenticated,
-    paths: web::Path<(HubuumClassID, HubuumObjectID, HubuumClassID, HubuumObjectID)>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    let user = &requestor.principal;
-    let (from_class, from_object, to_class, to_object) = paths.into_inner();
-
-    debug!(
-        message = "Creating object relation",
-        user_id = user.id(),
-        from_class = from_class.id(),
-        from_object = from_object.id(),
-        to_class = to_class.id(),
-        to_object = to_object.id()
-    );
-
-    let is_related = from_class.direct_relation_to(&pool, &to_class).await?;
-
-    if is_related.is_none() {
-        debug!(
-            message = "Relation does not exist",
-            user_id = user.id(),
-            from_class = from_class.id(),
-            to_class = to_class.id()
-        );
-        return Err(ApiError::NotFound(format!(
-            "Class {} is not related to class {}",
-            from_class.id(),
-            to_class.id()
-        )));
-    }
-
-    let relation = is_related.expect("Relation should exist after is_none check");
-
-    let relation = NewHubuumObjectRelation {
-        class_relation_id: relation.id,
-        from_hubuum_object_id: from_object.id(),
-        to_hubuum_object_id: to_object.id(),
-    };
-
-    if pool.permission_backend().uses_sql_permission_store() {
-        can!(
-            &pool,
-            user,
-            requestor.scopes(),
-            [Permissions::CreateObjectRelation],
-            from_class,
-            to_class
-        );
-    } else {
-        let resource = relation.to_resource_ref(&pool).await?;
-        authorize_resources(
-            pool.permission_backend(),
-            &pool,
-            user,
-            requestor.scopes(),
-            vec![Permissions::CreateObjectRelation],
-            vec![resource],
-        )
-        .await?;
-    }
-
-    let event_context = requestor.event_context(&req);
-    let relation = relation.save(&pool, &event_context).await?;
-
-    let location = api_locations::object_relation(
-        from_class.id(),
-        from_object.id(),
-        to_class.id(),
-        to_object.id(),
-    )?;
-    Ok(ApiResponse::created(relation, location))
-}
-
-#[utoipa::path(
-    get,
-    path = "/api/v1/classes/{class_id}/history",
-    tag = "classes",
-    security(("bearer_auth" = [])),
-    params(("class_id" = i32, Path, description = "Class ID")),
-    responses(
-        (status = 200, description = "Class history", body = [HistoryResponse<HubuumClassHistory>]),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 403, description = "Forbidden", body = ApiErrorResponse),
-        (status = 404, description = "Class not found", body = ApiErrorResponse)
-    )
-)]
-#[get("/{class_id}/history")]
-async fn get_class_history(
-    pool: AppContext,
-    requestor: Authenticated,
-    class_id: web::Path<HubuumClassID>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    use crate::api::v1::handlers::history::{
-        HistoryResponse, can_read_deleted_history, resolve_actor_usernames,
-    };
-
-    let user = &requestor.principal;
-    let class_id = class_id.into_inner();
-    let (entity_id, require_history) = match class_id.instance(&pool).await {
-        Ok(instance) => {
-            can!(
-                &pool,
-                user,
-                requestor.scopes(),
-                [Permissions::ReadClass],
-                instance
-            );
-            (instance.id, false)
-        }
-        Err(ApiError::NotFound(_)) if can_read_deleted_history(&pool, &requestor).await? => {
-            (class_id.id(), true)
-        }
-        Err(err) => return Err(err),
-    };
-
-    let params = parse_query_parameter(req.query_string())?;
-    let search_params = prepare_db_pagination::<HubuumClassHistory>(&params)?;
-    let (rows, total_count) =
-        class_history_paginated_with_total_count(entity_id, &pool, &search_params).await?;
-    if require_history && rows.is_empty() && params.cursor.is_none() {
-        return Err(ApiError::NotFound(format!("class {entity_id} not found")));
-    }
-
-    let actor_ids = rows.iter().filter_map(|r| r.actor_id).collect();
-    let actor_map = resolve_actor_usernames(&pool, actor_ids).await?;
-
-    ApiResponse::mapped_paginated(rows, total_count, &params, move |rows| {
-        rows.into_iter()
-            .map(|row| {
-                let actor_username = row.actor_id.and_then(|aid| actor_map.get(&aid).cloned());
-                HistoryResponse {
-                    entry: row,
-                    actor_username,
-                }
-            })
-            .collect()
-    })
-}
-
-#[utoipa::path(
-    get,
-    path = "/api/v1/classes/{class_id}/history/as-of",
-    tag = "classes",
-    security(("bearer_auth" = [])),
-    params(
-        ("class_id" = i32, Path, description = "Class ID"),
-        ("at" = String, Query, description = "RFC3339 timestamp")
-    ),
-    responses(
-        (status = 200, description = "Class version at timestamp", body = HistoryResponse<HubuumClassHistory>),
-        (status = 400, description = "Bad request", body = ApiErrorResponse),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 403, description = "Forbidden", body = ApiErrorResponse),
-        (status = 404, description = "Class or version not found", body = ApiErrorResponse)
-    )
-)]
-#[get("/{class_id}/history/as-of")]
-async fn get_class_as_of(
-    pool: AppContext,
-    requestor: Authenticated,
-    class_id: web::Path<HubuumClassID>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    use crate::api::v1::handlers::history::{
-        HistoryResponse, can_read_deleted_history, parse_as_of, resolve_actor_usernames,
-    };
-
-    let user = &requestor.principal;
-    let class_id = class_id.into_inner();
-    let entity_id = match class_id.instance(&pool).await {
-        Ok(instance) => {
-            can!(
-                &pool,
-                user,
-                requestor.scopes(),
-                [Permissions::ReadClass],
-                instance
-            );
-            instance.id
-        }
-        Err(ApiError::NotFound(_)) if can_read_deleted_history(&pool, &requestor).await? => {
-            class_id.id()
-        }
-        Err(err) => return Err(err),
-    };
-
-    let at = parse_as_of(req.query_string())?;
-    let row = class_as_of(entity_id, at, &pool)
-        .await?
-        .ok_or_else(|| ApiError::NotFound(format!("no version of class {entity_id} at {at}")))?;
-
-    let actor_map = resolve_actor_usernames(&pool, row.actor_id.into_iter().collect()).await?;
-    let actor_username = row.actor_id.and_then(|aid| actor_map.get(&aid).cloned());
-    Ok(ApiResponse::ok(HistoryResponse {
-        entry: row,
-        actor_username,
-    }))
-}
-
-#[utoipa::path(
-    get,
-    path = "/api/v1/classes/{class_id}/{object_id}/history",
-    tag = "classes",
-    security(("bearer_auth" = [])),
-    params(
-        ("class_id" = i32, Path, description = "Class ID"),
-        ("object_id" = i32, Path, description = "Object ID")
-    ),
-    responses(
-        (status = 200, description = "Object history", body = [HistoryResponse<HubuumObjectHistory>]),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 403, description = "Forbidden", body = ApiErrorResponse),
-        (status = 404, description = "Class or object not found", body = ApiErrorResponse)
-    )
-)]
-#[get("/{class_id}/{object_id}/history")]
-async fn get_object_history(
-    pool: AppContext,
-    requestor: Authenticated,
-    paths: web::Path<(HubuumClassID, HubuumObjectID)>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    use crate::api::v1::handlers::history::{
-        HistoryResponse, can_read_deleted_history, resolve_actor_usernames,
-    };
-
-    let user = &requestor.principal;
-    let (class_id, object_id) = paths.into_inner();
-
-    let (entity_id, require_history) =
-        match check_if_object_in_class(&pool, &class_id, &object_id).await {
-            Ok(()) => {
-                let object = object_id.instance(&pool).await?;
-                can!(
-                    &pool,
-                    user,
-                    requestor.scopes(),
-                    [Permissions::ReadObject],
-                    object
-                );
-                (object.id, false)
-            }
-            Err(ApiError::NotFound(_)) if can_read_deleted_history(&pool, &requestor).await? => {
-                (object_id.id(), true)
-            }
-            Err(err) => return Err(err),
-        };
-
-    let params = parse_query_parameter(req.query_string())?;
-    let search_params = prepare_db_pagination::<HubuumObjectHistory>(&params)?;
-    let (rows, total_count) =
-        object_history_paginated_with_total_count(entity_id, class_id.id(), &pool, &search_params)
-            .await?;
-    if require_history && rows.is_empty() && params.cursor.is_none() {
-        return Err(ApiError::NotFound(format!("object {entity_id} not found")));
-    }
-
-    let actor_ids = rows.iter().filter_map(|r| r.actor_id).collect();
-    let actor_map = resolve_actor_usernames(&pool, actor_ids).await?;
-
-    ApiResponse::mapped_paginated(rows, total_count, &params, move |rows| {
-        rows.into_iter()
-            .map(|row| {
-                let actor_username = row.actor_id.and_then(|aid| actor_map.get(&aid).cloned());
-                HistoryResponse {
-                    entry: row,
-                    actor_username,
-                }
-            })
-            .collect()
-    })
-}
-
-#[utoipa::path(
-    get,
-    path = "/api/v1/classes/{class_id}/{object_id}/history/as-of",
-    tag = "classes",
-    security(("bearer_auth" = [])),
-    params(
-        ("class_id" = i32, Path, description = "Class ID"),
-        ("object_id" = i32, Path, description = "Object ID"),
-        ("at" = String, Query, description = "RFC3339 timestamp")
-    ),
-    responses(
-        (status = 200, description = "Object version at timestamp", body = HistoryResponse<HubuumObjectHistory>),
-        (status = 400, description = "Bad request", body = ApiErrorResponse),
-        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
-        (status = 403, description = "Forbidden", body = ApiErrorResponse),
-        (status = 404, description = "Class, object, or version not found", body = ApiErrorResponse)
-    )
-)]
-#[get("/{class_id}/{object_id}/history/as-of")]
-async fn get_object_as_of(
-    pool: AppContext,
-    requestor: Authenticated,
-    paths: web::Path<(HubuumClassID, HubuumObjectID)>,
-    req: HttpRequest,
-) -> Result<impl Responder, ApiError> {
-    use crate::api::v1::handlers::history::{
-        HistoryResponse, can_read_deleted_history, parse_as_of, resolve_actor_usernames,
-    };
-
-    let user = &requestor.principal;
-    let (class_id, object_id) = paths.into_inner();
-
-    let entity_id = match check_if_object_in_class(&pool, &class_id, &object_id).await {
-        Ok(()) => {
-            let object = object_id.instance(&pool).await?;
-            can!(
-                &pool,
-                user,
-                requestor.scopes(),
-                [Permissions::ReadObject],
-                object
-            );
-            object.id
-        }
-        Err(ApiError::NotFound(_)) if can_read_deleted_history(&pool, &requestor).await? => {
-            object_id.id()
-        }
-        Err(err) => return Err(err),
-    };
-
-    let at = parse_as_of(req.query_string())?;
-    let row = object_as_of(entity_id, class_id.id(), at, &pool)
-        .await?
-        .ok_or_else(|| ApiError::NotFound(format!("no version of object {entity_id} at {at}")))?;
-
-    let actor_map = resolve_actor_usernames(&pool, row.actor_id.into_iter().collect()).await?;
-    let actor_username = row.actor_id.and_then(|aid| actor_map.get(&aid).cloned());
-    Ok(ApiResponse::ok(HistoryResponse {
-        entry: row,
-        actor_username,
-    }))
-}
+mod class_objects;
+mod class_related;
+mod computed_objects;
+mod history_endpoints;
+mod object_related;
+
+pub use class_objects::*;
+pub use class_related::*;
+pub use history_endpoints::*;
+pub use object_related::*;
