+++ src/models/mod.rs
@@ -11,6 +11,7 @@ pub mod group;
 pub mod identity;
 pub mod import;
 pub mod object;
+pub mod object_data_patch;
 pub mod output;
 pub mod permissions;
 pub mod principal;
@@ -38,6 +39,7 @@ pub use crate::models::group::*;
 pub use crate::models::identity::*;
 pub use crate::models::import::*;
 pub use crate::models::object::*;
+pub use crate::models::object_data_patch::*;
 pub use crate::models::output::*;
 pub use crate::models::permissions::*;
 pub use crate::models::principal::*;
